+++ nrfx/drivers/src/nrfx_pwm.c
@@ -75,7 +75,7 @@ typedef struct
     nrfx_pwm_handler_t        handler;
     void *                    p_context;
     nrfx_drv_state_t volatile state;
-    uint8_t                   flags;
+    uint32_t                  flags;
     bool                      skip_gpio_cfg;
 } pwm_control_block_t;
 static pwm_control_block_t m_cb[NRFX_PWM_ENABLED_COUNT];
@@ -194,7 +194,11 @@ nrfx_err_t nrfx_pwm_init(nrfx_pwm_t const *        p_instance,
 
     if (p_cb->state != NRFX_DRV_STATE_UNINITIALIZED)
     {
+#if NRFX_API_VER_AT_LEAST(3, 2, 0)
+        err_code = NRFX_ERROR_ALREADY;
+#else
         err_code = NRFX_ERROR_INVALID_STATE;
+#endif
         NRFX_LOG_WARNING("Function: %s, error code: %s.",
                          __func__,
                          NRFX_LOG_ERROR_STRING_GET(err_code));
@@ -221,8 +225,9 @@ nrfx_err_t nrfx_pwm_init(nrfx_pwm_t const *        p_instance,
 
 nrfx_err_t nrfx_pwm_reconfigure(nrfx_pwm_t const * p_instance, nrfx_pwm_config_t const * p_config)
 {
-    NRFX_ASSERT(p_config);
     pwm_control_block_t * p_cb = &m_cb[p_instance->instance_id];
+
+    NRFX_ASSERT(p_config);
 
     if (p_cb->state == NRFX_DRV_STATE_UNINITIALIZED)
     {
@@ -241,6 +246,7 @@ nrfx_err_t nrfx_pwm_reconfigure(nrfx_pwm_t const * p_instance, nrfx_pwm_config_t
 void nrfx_pwm_uninit(nrfx_pwm_t const * p_instance)
 {
     pwm_control_block_t * p_cb = &m_cb[p_instance->instance_id];
+
     NRFX_ASSERT(p_cb->state != NRFX_DRV_STATE_UNINITIALIZED);
 
     nrfy_pwm_int_uninit(p_instance->p_reg);
@@ -256,11 +262,19 @@ void nrfx_pwm_uninit(nrfx_pwm_t const * p_instance)
     }
 
     p_cb->state = NRFX_DRV_STATE_UNINITIALIZED;
+    NRFX_LOG_INFO("Uninitialized.");
+}
+
+bool nrfx_pwm_init_check(nrfx_pwm_t const * p_instance)
+{
+    pwm_control_block_t * p_cb = &m_cb[p_instance->instance_id];
+
+    return (p_cb->state != NRFX_DRV_STATE_UNINITIALIZED);
 }
 
 static uint32_t start_playback(nrfx_pwm_t const *    p_instance,
                                pwm_control_block_t * p_cb,
-                               uint8_t               flags,
+                               uint32_t              flags,
                                uint8_t               seq_id)
 {
     p_cb->state = NRFX_DRV_STATE_POWERED_ON;
@@ -294,7 +308,7 @@ static uint32_t start_playback(nrfx_pwm_t const *    p_instance,
 #endif
         if (flags & NRFX_PWM_FLAG_NO_EVT_FINISHED)
         {
-            int_mask &= ~NRF_PWM_INT_LOOPSDONE_MASK;
+            int_mask &= (uint32_t)~NRF_PWM_INT_LOOPSDONE_MASK;
         }
 
         nrfy_pwm_int_set(p_instance->p_reg, int_mask);
@@ -334,6 +348,7 @@ uint32_t nrfx_pwm_simple_playback(nrfx_pwm_t const *         p_instance,
                                   uint32_t                   flags)
 {
     pwm_control_block_t * p_cb = &m_cb[p_instance->instance_id];
+
     NRFX_ASSERT(p_cb->state != NRFX_DRV_STATE_UNINITIALIZED);
     NRFX_ASSERT(playback_count > 0);
     NRFX_ASSERT(nrfx_is_in_ram(p_sequence->values.p_raw));
@@ -344,7 +359,7 @@ uint32_t nrfx_pwm_simple_playback(nrfx_pwm_t const *         p_instance,
     nrfy_pwm_sequence_set(p_instance->p_reg, 1, p_sequence);
     bool odd = (playback_count & 1);
     nrfy_pwm_loop_set(p_instance->p_reg,
-        (playback_count / 2) + (odd ? 1 : 0));
+        (uint16_t)((playback_count / 2UL) + (odd ? 1UL : 0UL)));
 
     uint32_t shorts_mask = 0;
     if (flags & NRFX_PWM_FLAG_STOP)
@@ -378,6 +393,7 @@ uint32_t nrfx_pwm_complex_playback(nrfx_pwm_t const *         p_instance,
                                    uint32_t                   flags)
 {
     pwm_control_block_t * p_cb = &m_cb[p_instance->instance_id];
+
     NRFX_ASSERT(p_cb->state != NRFX_DRV_STATE_UNINITIALIZED);
     NRFX_ASSERT(playback_count > 0);
     NRFX_ASSERT(nrfx_is_in_ram(p_sequence_0->values.p_raw));
@@ -420,6 +436,7 @@ uint32_t nrfx_pwm_complex_playback(nrfx_pwm_t const *         p_instance,
 bool nrfx_pwm_stop(nrfx_pwm_t const * p_instance, bool wait_until_stopped)
 {
     pwm_control_block_t * p_cb = &m_cb[p_instance->instance_id];
+
     NRFX_ASSERT(p_cb->state != NRFX_DRV_STATE_UNINITIALIZED);
 
     bool ret_val = false;