+++ nrfx/samples/src/nrfx_gppi/fork/main.c
@@ -100,10 +100,9 @@ int main(void)
     uint8_t gppi_channel;
 
 #if defined(__ZEPHYR__)
-    IRQ_DIRECT_CONNECT(NRFX_IRQ_NUMBER_GET(NRF_TIMER_INST_GET(TIMER_INST_IDX)), IRQ_PRIO_LOWEST,
-                       NRFX_TIMER_INST_HANDLER_GET(TIMER_INST_IDX), 0);
-    IRQ_DIRECT_CONNECT(NRFX_IRQ_NUMBER_GET(NRF_GPIOTE), IRQ_PRIO_LOWEST, nrfx_gpiote_irq_handler,
-                       0);
+    IRQ_CONNECT(NRFX_IRQ_NUMBER_GET(NRF_TIMER_INST_GET(TIMER_INST_IDX)), IRQ_PRIO_LOWEST,
+                NRFX_TIMER_INST_HANDLER_GET(TIMER_INST_IDX), 0, 0);
+    IRQ_CONNECT(NRFX_IRQ_NUMBER_GET(NRF_GPIOTE), IRQ_PRIO_LOWEST, nrfx_gpiote_irq_handler, 0, 0);
 #endif
 
     NRFX_EXAMPLE_LOG_INIT();
@@ -113,7 +112,8 @@ int main(void)
 
     status = nrfx_gpiote_init(NRFX_GPIOTE_DEFAULT_CONFIG_IRQ_PRIORITY);
     NRFX_ASSERT(status == NRFX_SUCCESS);
-    NRFX_LOG_INFO("GPIOTE status: %s", nrfx_gpiote_is_init() ? "initialized" : "not initialized");
+    NRFX_LOG_INFO("GPIOTE status: %s",
+                  nrfx_gpiote_is_init() ? "initialized" : "not initialized");
 
     status = nrfx_gpiote_channel_alloc(&out_channel_primary);
     NRFX_ASSERT(status == NRFX_SUCCESS);
@@ -139,9 +139,7 @@ int main(void)
         .init_val = NRF_GPIOTE_INITIAL_VALUE_HIGH,
     };
 
-    status = nrfx_gpiote_output_configure(OUTPUT_PIN_PRIMARY,
-                                          &output_config,
-                                          &task_config);
+    status = nrfx_gpiote_output_configure(OUTPUT_PIN_PRIMARY, &output_config, &task_config);
     NRFX_ASSERT(status == NRFX_SUCCESS);
 
     nrfx_gpiote_out_task_enable(OUTPUT_PIN_PRIMARY);
@@ -153,10 +151,7 @@ int main(void)
     task_config.task_ch = out_channel_fork;
     task_config.init_val = NRF_GPIOTE_INITIAL_VALUE_LOW;
 
-    status = nrfx_gpiote_output_configure(
-        OUTPUT_PIN_FORK,
-        &output_config,
-        &task_config);
+    status = nrfx_gpiote_output_configure(OUTPUT_PIN_FORK, &output_config, &task_config);
     NRFX_ASSERT(status == NRFX_SUCCESS);
 
     nrfx_gpiote_out_task_enable(OUTPUT_PIN_FORK);