+++ nrfx/samples/src/nrfx_saadc/common/saadc_examples_common.c
@@ -65,8 +65,7 @@ void gpiote_pin_toggle_task_setup(nrfx_gpiote_pin_t pin)
     nrfx_gpiote_out_task_enable(pin);
 }
 
-void pin_on_event_toggle_setup(nrfx_gpiote_pin_t pin,
-                               uint32_t eep)
+void pin_on_event_toggle_setup(nrfx_gpiote_pin_t pin, uint32_t eep)
 {
     nrfx_err_t status;
     (void)status;
@@ -78,7 +77,9 @@ void pin_on_event_toggle_setup(nrfx_gpiote_pin_t pin,
     status = nrfx_gppi_channel_alloc(&gppi_channel);
     NRFX_ASSERT(status == NRFX_SUCCESS);
 
-    nrfx_gppi_channel_endpoints_setup(gppi_channel, eep, nrfx_gpiote_out_task_address_get(pin));
+    nrfx_gppi_channel_endpoints_setup(gppi_channel,
+                                      eep,
+                                      nrfx_gpiote_out_task_address_get(pin));
 
     nrfx_gppi_channels_enable(NRFX_BIT(gppi_channel));
 }