+++ nrfx/samples/src/nrfx_saadc/simple_non_blocking/main.c
@@ -97,14 +97,21 @@ static const nrfx_saadc_channel_t m_multiple_channels[] =
 /** @brief Symbol specifying numbers of multiple channels ( @ref m_multiple_channels) used by SAADC. */
 #define CHANNEL_COUNT NRFX_ARRAY_SIZE(m_multiple_channels)
 
+/** @brief Symbol specifying the number of SAADC samplings to trigger. */
+#define SAMPLING_ITERATIONS 8
+
+/** @brief Symbol specifying the resolution of the SAADC. */
+#define RESOLUTION NRF_SAADC_RESOLUTION_8BIT
+
 /** @brief Array specifying GPIO pins used to test the functionality of SAADC. */
 static uint8_t m_out_pins[CHANNEL_COUNT] = {LOOPBACK_PIN_1B, LOOPBACK_PIN_2B, LOOPBACK_PIN_3B};
 
 /** @brief Samples buffer defined with the size of @ref CHANNEL_COUNT symbol to store values from each channel ( @ref m_multiple_channels). */
-static nrf_saadc_value_t m_samples_buffer[CHANNEL_COUNT];
-
-/** @brief Symbol specifying the number of SAADC samplings to trigger. */
-#define SAMPLING_ITERATIONS 8
+#if (NRF_SAADC_8BIT_SAMPLE_WIDTH == 8) && (RESOLUTION == NRF_SAADC_RESOLUTION_8BIT)
+static uint8_t m_samples_buffer[CHANNEL_COUNT];
+#else
+static uint16_t m_samples_buffer[CHANNEL_COUNT];
+#endif
 
 /** @brief Enum with the current state of the simple state machine. */
 static state_t m_current_state = STATE_SINGLE_CONFIG;
@@ -132,7 +139,8 @@ static void saadc_handler(nrfx_saadc_evt_t const * p_event)
             samples_number = p_event->data.done.size;
             for (uint16_t i = 0; i < samples_number; i++)
             {
-                NRFX_LOG_INFO("[Sample %d] value == %d", i, p_event->data.done.p_buffer[i]);
+                NRFX_LOG_INFO("[Sample %d] value == %d",
+                              i, NRFX_SAADC_SAMPLE_GET(RESOLUTION, p_event->data.done.p_buffer, i));
             }
 
             m_saadc_ready = true;
@@ -159,6 +167,11 @@ int main(void)
     nrfx_err_t status;
     (void)status;
 
+#if defined(__ZEPHYR__)
+    IRQ_CONNECT(NRFX_IRQ_NUMBER_GET(NRF_SAADC), IRQ_PRIO_LOWEST, nrfx_saadc_irq_handler, 0, 0);
+    IRQ_CONNECT(NRFX_IRQ_NUMBER_GET(NRF_GPIOTE), IRQ_PRIO_LOWEST, nrfx_gpiote_irq_handler, 0, 0);
+#endif
+
     NRFX_EXAMPLE_LOG_INIT();
     NRFX_LOG_INFO("Starting nrfx_saadc simple non-blocking example.");
     NRFX_EXAMPLE_LOG_PROCESS();
@@ -166,15 +179,16 @@ int main(void)
     status = nrfx_saadc_init(NRFX_SAADC_DEFAULT_CONFIG_IRQ_PRIORITY);
     NRFX_ASSERT(status == NRFX_SUCCESS);
 
+    status = nrfx_gpiote_init(NRFX_GPIOTE_DEFAULT_CONFIG_IRQ_PRIORITY);
+    NRFX_ASSERT(status == NRFX_SUCCESS);
+    NRFX_LOG_INFO("GPIOTE status: %s",
+                  nrfx_gpiote_is_init() ? "initialized" : "not initialized");
+
     uint8_t i;
     for (i = 0; i < CHANNEL_COUNT; i++)
     {
         gpiote_pin_toggle_task_setup(m_out_pins[i]);
     }
-
-#if defined(__ZEPHYR__)
-    IRQ_DIRECT_CONNECT(NRFX_IRQ_NUMBER_GET(NRF_SAADC), IRQ_PRIO_LOWEST, nrfx_saadc_irq_handler, 0);
-#endif
 
     uint32_t sampling_index = 0;
     while (1)
@@ -189,7 +203,7 @@ int main(void)
 
                 uint32_t channels_mask = nrfx_saadc_channels_configured_get();
                 status = nrfx_saadc_simple_mode_set(channels_mask,
-                                                    NRF_SAADC_RESOLUTION_8BIT,
+                                                    RESOLUTION,
                                                     NRF_SAADC_OVERSAMPLE_DISABLED,
                                                     saadc_handler);
                 NRFX_ASSERT(status == NRFX_SUCCESS);