+++ nrfx/samples/src/nrfx_twim_twis/tx_rx_non_blocking/main.c
@@ -187,6 +187,14 @@ int main(void)
     nrfx_err_t status;
     (void)status;
 
+#if defined(__ZEPHYR__)
+    IRQ_CONNECT(NRFX_IRQ_NUMBER_GET(NRF_TWIM_INST_GET(TWIM_INST_IDX)), IRQ_PRIO_LOWEST,
+                NRFX_TWIM_INST_HANDLER_GET(TWIM_INST_IDX), 0, 0);
+
+    IRQ_CONNECT(NRFX_IRQ_NUMBER_GET(NRF_TWIS_INST_GET(TWIS_INST_IDX)), IRQ_PRIO_LOWEST,
+                NRFX_TWIS_INST_HANDLER_GET(TWIS_INST_IDX), 0, 0);
+#endif
+
     NRFX_EXAMPLE_LOG_INIT();
 
     NRFX_LOG_INFO("Starting nrfx_twim_twis non-blocking example.");
@@ -204,14 +212,6 @@ int main(void)
                                                               SLAVE_ADDR);
     status = nrfx_twis_init(&twis_inst, &twis_config, twis_handler);
     NRFX_ASSERT(status == NRFX_SUCCESS);
-
-#if defined(__ZEPHYR__)
-    IRQ_DIRECT_CONNECT(NRFX_IRQ_NUMBER_GET(NRF_TWIM_INST_GET(TWIM_INST_IDX)), IRQ_PRIO_LOWEST,
-                       NRFX_TWIM_INST_HANDLER_GET(TWIM_INST_IDX), 0);
-
-    IRQ_DIRECT_CONNECT(NRFX_IRQ_NUMBER_GET(NRF_TWIS_INST_GET(TWIS_INST_IDX)), IRQ_PRIO_LOWEST,
-                       NRFX_TWIS_INST_HANDLER_GET(TWIS_INST_IDX), 0);
-#endif
 
     NRFX_LOG_INFO("Message to send: %s", m_tx_buffer_master);
     NRFX_LOG_INFO("Content of master RX buffer: %s", m_rx_buffer_master);